// web/src/app/alarm/types/monitor.ts
export interface SearchParams {
    time?: number;
    end?: number;
    start?: number;
    step?: number;
    query: string;
}